+++ packages/ragbits-document-search/src/ragbits/document_search/retrieval/rerankers/rerankers_answerdotai.py
@@ -62,6 +62,7 @@ class AnswerAIReranker(Reranker[RerankerOptions]):
             )
             if merged_options.top_n:
                 response = response.top_k(merged_options.top_n)
+
             outputs.results = [element_list[result.document.doc_id] for result in response]
 
             return outputs.results