+++ packages/ragbits-document-search/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Add optional score override on reranking step (#544)
+- Add score threshold to reranker options (#544)
 - Add LLM reranker (#513)
 
 ## 0.16.0 (2025-04-29)